+++ projections/typescript/index.ts
@@ -1,0 +1,2 @@
+export * from "./capsule.js";
+export * from "./validator-api.js";